# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/{provider}": {
      "get": {
        "operationId": "AuthController_startOAuth",
        "parameters": [],
        "responses": {
          "302": {
            "description": "Redirect to provider for authentication"
          },
          "400": {
            "description": "Provider not found or disabled"
          }
        },
        "summary": "Start OAuth flow for provider",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/access-tokens/{tokenId}": {
      "delete": {
        "operationId": "AccessTokenController_revokeAccessToken",
        "parameters": [
          {
            "name": "tokenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access token revoked successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a specific access token",
        "tags": [
          "Access Tokens"
        ]
      }
    },
...
    "/api/v1/oauth-providers": {
      "post": {
        "operationId": "OAuthProvidersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOAuthProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OAuth provider created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthProvider"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "409": {
            "description": "Provider with same providerId already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new OAuth provider (Admin only)",
        "tags": [
          "OAuth Providers"
        ]
      },
      "get": {
        "operationId": "OAuthProvidersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all OAuth providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OAuthProvider"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all OAuth providers (Admin only)",
        "tags": [
          "OAuth Providers"
        ]
      }
    },
    "/api/v1/oauth-providers/{id}": {
      "get": {
        "operationId": "OAuthProvidersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "OAuth provider ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth provider found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthProvider"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "OAuth provider not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get OAuth provider by ID (Admin only)",
        "tags": [
          "OAuth Providers"
        ]
      },
      "patch": {
        "operationId": "OAuthProvidersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "OAuth provider ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOAuthProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth provider updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthProvider"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "OAuth provider not found"
          },
          "409": {
            "description": "Provider with same providerId already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update OAuth provider (Admin only)",
        "tags": [
          "OAuth Providers"
        ]
      },
      "delete": {
        "operationId": "OAuthProvidersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "OAuth provider ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OAuth provider deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "OAuth provider not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete OAuth provider (Admin only)",
        "tags": [
          "OAuth Providers"
        ]
      }
    },
    "/api/v1/oauth-providers/by-provider/{providerId}": {
      "get": {
        "operationId": "OAuthProvidersController_findByProviderId",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "OAuth provider identifier (e.g., github, google)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth provider found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthProvider"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "OAuth provider not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get OAuth provider by provider ID (Admin only)",
        "tags": [
          "OAuth Providers"
        ]
      }
    },
    "/api/v1/oauth-providers/{id}/toggle": {
      "patch": {
        "operationId": "OAuthProvidersController_toggleEnabled",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "OAuth provider ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth provider status toggled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthProvider"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "OAuth provider not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle OAuth provider enabled status (Admin only)",
        "tags": [
          "OAuth Providers"
        ]
      }
    },
    "/api/v1/oauth-providers/{id}/config": {
      "get": {
        "operationId": "OAuthProvidersController_getProviderConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "OAuth provider ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth provider configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clientId": {
                      "type": "string"
                    },
                    "clientSecret": {
                      "type": "string"
                    },
                    "callbackUrl": {
                      "type": "string"
                    },
                    "scopes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "GITHUB",
                        "GOOGLE",
                        "CUSTOM"
                      ]
                    },
                    "authorizationUrl": {
                      "type": "string"
                    },
                    "tokenUrl": {
                      "type": "string"
                    },
                    "userInfoUrl": {
                      "type": "string"
                    },
                    "profileFields": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "additionalConfig": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "OAuth provider not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get OAuth provider configuration for authentication (Admin only)",
        "tags": [
          "OAuth Providers"
        ]
      }
    },
...
    "/api/v1/system-access-tokens": {
      "post": {
        "operationId": "SystemAccessTokenController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create system access token",
        "tags": [
          "System Access Tokens"
        ]
      },
      "get": {
        "operationId": "SystemAccessTokenController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "System Access Tokens"
        ]
      }
    },
    "/api/v1/system-access-tokens/{id}": {
      "delete": {
        "operationId": "SystemAccessTokenController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "System Access Tokens"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zentik API",
    "description": "The Zentik notification API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OAuthProviderType": {
        "type": "string",
        "enum": [
          "GITHUB",
          "GOOGLE",
          "CUSTOM"
        ],
        "description": "The type of provider (GitHub, Google, or custom)"
      },
      "CreateOAuthProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the OAuth provider"
          },
          "providerId": {
            "type": "string",
            "description": "The unique identifier for the provider (e.g., github, google, custom)"
          },
          "type": {
            "description": "The type of provider (GitHub, Google, or custom)",
            "allOf": [
              {
                "$ref": "#/components/schemas/OAuthProviderType"
              }
            ]
          },
          "clientId": {
            "type": "string",
            "description": "The OAuth client ID"
          },
          "clientSecret": {
            "type": "string",
            "description": "The OAuth client secret"
          },
          "callbackUrl": {
            "type": "string",
            "description": "Custom callback URL (if not using default)"
          },
          "scopes": {
            "description": "Array of OAuth scopes",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether the provider is currently enabled"
          },
          "iconUrl": {
            "type": "string",
            "description": "Icon URL for the provider"
          },
          "color": {
            "type": "string",
            "description": "Hex color code for the provider (e.g., #FF5733)",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#0a7ea4"
          },
          "textColor": {
            "type": "string",
            "description": "Hex color code for the provider text (e.g., #FFFFFF)",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#FFFFFF"
          },
          "authorizationUrl": {
            "type": "string",
            "description": "Authorization URL for custom OAuth provider"
          },
          "tokenUrl": {
            "type": "string",
            "description": "Token URL for custom OAuth provider"
          },
          "userInfoUrl": {
            "type": "string",
            "description": "User info URL for custom OAuth provider"
          },
          "profileFields": {
            "description": "Profile fields mapping for custom OAuth provider",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "additionalConfig": {
            "type": "string",
            "description": "Additional configuration as JSON string"
          }
        },
        "required": [
          "name",
          "providerId",
          "type",
          "clientId",
          "clientSecret",
          "scopes"
        ]
      },
      "OAuthProvider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "description": "The name of the OAuth provider"
          },
          "providerId": {
            "type": "string",
            "description": "The unique identifier for the provider (e.g., github, google, custom)"
          },
          "type": {
            "description": "The type of provider (GitHub, Google, or custom)",
            "allOf": [
              {
                "$ref": "#/components/schemas/OAuthProviderType"
              }
            ]
          },
          "clientId": {
            "type": "string",
            "description": "The OAuth client ID"
          },
          "clientSecret": {
            "type": "string",
            "description": "The OAuth client secret"
          },
          "callbackUrl": {
            "type": "string",
            "description": "Custom callback URL (if not using default)"
          },
          "scopes": {
            "description": "Array of OAuth scopes",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether the provider is currently enabled"
          },
          "iconUrl": {
            "type": "string",
            "description": "Icon URL for the provider"
          },
          "color": {
            "type": "string",
            "description": "Hex color code for the provider (e.g., #FF5733)",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#0a7ea4"
          },
          "textColor": {
            "type": "string",
            "description": "Hex color code for the provider text (e.g., #FFFFFF)",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#FFFFFF"
          },
          "authorizationUrl": {
            "type": "string",
            "description": "Authorization URL for custom OAuth provider"
          },
          "tokenUrl": {
            "type": "string",
            "description": "Token URL for custom OAuth provider"
          },
          "userInfoUrl": {
            "type": "string",
            "description": "User info URL for custom OAuth provider"
          },
          "profileFields": {
            "description": "Profile fields mapping for custom OAuth provider",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "additionalConfig": {
            "type": "string",
            "description": "Additional configuration as JSON string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "providerId",
          "type",
          "clientId",
          "clientSecret",
          "callbackUrl",
          "scopes",
          "isEnabled",
          "iconUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateOAuthProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the OAuth provider"
          },
          "providerId": {
            "type": "string",
            "description": "The unique identifier for the provider (e.g., github, google, custom)"
          },
          "type": {
            "description": "The type of provider (GitHub, Google, or custom)",
            "allOf": [
              {
                "$ref": "#/components/schemas/OAuthProviderType"
              }
            ]
          },
          "clientId": {
            "type": "string",
            "description": "The OAuth client ID"
          },
          "clientSecret": {
            "type": "string",
            "description": "The OAuth client secret"
          },
          "callbackUrl": {
            "type": "string",
            "description": "Custom callback URL (if not using default)"
          },
          "scopes": {
            "description": "Array of OAuth scopes",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether the provider is currently enabled"
          },
          "iconUrl": {
            "type": "string",
            "description": "Icon URL for the provider"
          },
          "color": {
            "type": "string",
            "description": "Hex color code for the provider (e.g., #FF5733)",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#0a7ea4"
          },
          "textColor": {
            "type": "string",
            "description": "Hex color code for the provider text (e.g., #FFFFFF)",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#FFFFFF"
          },
          "authorizationUrl": {
            "type": "string",
            "description": "Authorization URL for custom OAuth provider"
          },
          "tokenUrl": {
            "type": "string",
            "description": "Token URL for custom OAuth provider"
          },
          "userInfoUrl": {
            "type": "string",
            "description": "User info URL for custom OAuth provider"
          },
          "profileFields": {
            "description": "Profile fields mapping for custom OAuth provider",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "additionalConfig": {
            "type": "string",
            "description": "Additional configuration as JSON string"
          }
        }
      },
...
    }
  }
}